Fix some MSVC warnings.

+++ src/shared/apis/graphics/vulkan/instance.cpp
@@ -202,7 +202,7 @@ namespace pbr::shared::apis::graphics::vulkan {
         enables[0] = VK_VALIDATION_FEATURE_ENABLE_BEST_PRACTICES_EXT;
 
         features.sType = VK_STRUCTURE_TYPE_VALIDATION_FEATURES_EXT;
-        features.enabledValidationFeatureCount = enables.size();
+        features.enabledValidationFeatureCount = static_cast<uint32_t>(enables.size());
         features.pEnabledValidationFeatures = enables.data();
 
         info.pNext = &features;
